+++ cmd/vault2env.go
@@ -17,7 +17,7 @@ func main() {
 	config, err := config.Load(configPath)
 
 	if err != nil {
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
 
@@ -54,9 +54,9 @@ func run(command string, config *config.Config) error {
 }
 
 func showHelp() {
-	fmt.Println("Usage: vault2env [--config=<config_file>] <command>")
-	fmt.Println("Commands:")
-	fmt.Println("  pull: Pull secrets from Vault and write to .env")
-	fmt.Println("  push: Push secrets from .env to Vault")
-	fmt.Println("  help: Show this help")
+	fmt.Fprintln(os.Stderr, "Usage: vault2env [--config=<config_file>] <command>")
+	fmt.Fprintln(os.Stderr, "Commands:")
+	fmt.Fprintln(os.Stderr, "  pull: Pull secrets from Vault and write to .env")
+	fmt.Fprintln(os.Stderr, "  push: Push secrets from .env to Vault")
+	fmt.Fprintln(os.Stderr, "  help: Show this help")
 }