+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   JSON schema definition for config file.
+-   Input and output can be piped.
 
 ### Changed